Transliteration minor bug fix
